Extract TlvChunkParser to its own file and add test (#308)
- Extract TlvChunkParser from SctpReactorSpine.kt to its own minimal implementation file TlvChunkParser.kt in borg.trikeshed.sctp.
- Update imports in SctpReactorSpine.kt.
- Add TlvChunkParseTest.kt to test parsing TLV chunks and unknown-skip behavior based on chunk action bits.
