+++ src/commonMain/kotlin/borg/trikeshed/reactor/ngsctp/SctpReactorSpine.kt
@@ -20,6 +20,7 @@ import borg.trikeshed.context.nuid.Subnet
 import borg.trikeshed.context.nuid.NuidFanoutElement
 import borg.trikeshed.context.StreamHandle
 import borg.trikeshed.sctp.SctpElement
+import borg.trikeshed.sctp.TlvChunkParser
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -27,37 +28,6 @@ import kotlinx.coroutines.cancel
 import kotlinx.coroutines.channels.Channel
 
 // 1. TLV Chunk Parser
-class TlvChunkParser {
-    class Chunk(val type: Int, val data: ByteArray)
-
-    fun parse(data: ByteArray): Series<Chunk> {
-        val resultList = mutableListOf<Chunk>()
-        var offset = 0
-        while (offset < data.size) {
-            if (offset + 4 > data.size) break
-            val type = data[offset].toInt() and 0xFF
-            val length = ((data[offset + 2].toInt() and 0xFF) shl 8) or (data[offset + 3].toInt() and 0xFF)
-            if (offset + length > data.size) break
-
-            val chunkData = data.copyOfRange(offset + 4, offset + length)
-
-            if (type == 0x00) {
-                resultList.add(Chunk(type, chunkData))
-            } else {
-                val action = type shr 6
-                if (action == 0 || action == 1) {
-                    break // Stop processing
-                }
-                // skip others implicitly
-            }
-
-            val padding = (4 - (length % 4)) % 4
-            offset += length + padding
-        }
-
-        return resultList.size j { i -> resultList[i] }
-    }
-}
 
 // 2. Bounded Channel Stream
 class BoundedChannelStream(capacity: Int) {
